# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MealSync API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  owner: "owner",
  ngo: "ngo",
  resident: "resident",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  pgName?: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  owner: "owner",
  ngo: "ngo",
  resident: "resident",
} as const;

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
  role: RegisterBodyRole;
  pgName?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export type MealStatus = (typeof MealStatus)[keyof typeof MealStatus];

export const MealStatus = {
  pending: "pending",
  served: "served",
  completed: "completed",
  draft: "draft",
  published: "published",
} as const;

export interface Meal {
  id: number;
  menu?: string;
  breakfastMenu?: string;
  breakfastTime?: string;
  lunchMenu?: string;
  dinnerMenu?: string;
  lunchTime?: string;
  dinnerTime?: string;
  notes?: string;
  expectedPeople: number;
  predictedMeals?: number;
  actualServed?: number;
  leftoverMeals?: number;
  date: string;
  ngoNotified?: boolean;
  status: MealStatus;
}

export type UpsertMenuBodyStatus =
  (typeof UpsertMenuBodyStatus)[keyof typeof UpsertMenuBodyStatus];

export const UpsertMenuBodyStatus = {
  draft: "draft",
  published: "published",
} as const;

export interface UpsertMenuBody {
  breakfastMenu?: string;
  breakfastTime?: string;
  lunchMenu: string;
  dinnerMenu: string;
  lunchTime: string;
  dinnerTime: string;
  expectedPeople: number;
  notes?: string;
  status: UpsertMenuBodyStatus;
}

export interface CreateMealBody {
  menu: string;
  expectedPeople: number;
}

export interface MealPrediction {
  predictedMeals: number;
  confidence: number;
  message: string;
}

export interface LeftoverBody {
  leftoverMeals: number;
  actualServed: number;
}

export type NgoRequestStatus =
  (typeof NgoRequestStatus)[keyof typeof NgoRequestStatus];

export const NgoRequestStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
  completed: "completed",
} as const;

export interface NgoRequest {
  id: number;
  pgName: string;
  pgLocation: string;
  availableMeals: number;
  pickupTime: string;
  status: NgoRequestStatus;
  createdAt: string;
  mealMenu?: string;
}

export interface ConfirmMealBody {
  willEat: boolean;
  mealDate: string;
}

export interface MealConfirmation {
  id: number;
  willEat: boolean;
  mealDate: string;
  message: string;
}

export interface FeedbackBody {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string;
  mealDate: string;
}

export interface Feedback {
  id: number;
  rating: number;
  comment?: string;
  mealDate: string;
  message: string;
}

export interface OwnerStats {
  totalMealsToday: number;
  predictedMeals: number;
  leftoverMeals: number;
  mealsServed: number;
}

export interface GreenScore {
  score: number;
  mealsSavedToday: number;
  totalMealsSaved: number;
  message: string;
}

export interface DailyTrend {
  date: string;
  predicted: number;
  actual: number;
  leftover: number;
  wasteReduction: number;
}