# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MealSync API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meals:
    get:
      operationId: listMeals
      tags: [meals]
      summary: List all meals for the logged-in PG
      responses:
        "200":
          description: List of meals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meal"
    post:
      operationId: createMeal
      tags: [meals]
      summary: Create a meal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMealBody"
      responses:
        "201":
          description: Meal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
  /meals/today:
    get:
      operationId: getTodaysMenu
      tags: [meals]
      summary: Get today's menu for the logged-in PG owner
      responses:
        "200":
          description: Today's menu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
        "404":
          description: No menu found for today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTodaysMenu
      tags: [meals]
      summary: Delete today's menu if it's in draft status
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Cannot delete published menu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No menu found for today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /meals/upsert:
    post:
      operationId: upsertTodaysMenu
      tags: [meals]
      summary: Create or update today's menu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertMenuBody"
      responses:
        "200":
          description: Menu saved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
  /meals/{id}/predict:
    post:
      operationId: predictMeal
      tags: [meals]
      summary: Predict meals for a given meal entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Prediction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealPrediction"
  /meals/{id}/leftover:
    post:
      operationId: reportLeftover
      tags: [meals]
      summary: Report leftover meals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeftoverBody"
      responses:
        "200":
          description: Leftover reported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
  /meals/{id}/notify-ngo:
    post:
      operationId: notifyNgo
      tags: [meals]
      summary: Notify NGO about available leftovers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: NGO notified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NgoRequest"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Meal:
      type: object
      properties:
        id:
          type: integer
        menu:
          type: string
        breakfastMenu:
          type: string
        breakfastTime:
          type: string
        lunchMenu:
          type: string
        dinnerMenu:
          type: string
        lunchTime:
          type: string
        dinnerTime:
          type: string
        notes:
          type: string
        expectedPeople:
          type: integer
        predictedMeals:
          type: integer
        actualServed:
          type: integer
        leftoverMeals:
          type: integer
        date:
          type: string
        ngoNotified:
          type: boolean
        status:
          type: string
          enum: [pending, served, completed, draft, published]
      required:
        - id
        - expectedPeople
        - date
        - status
    UpsertMenuBody:
      type: object
      properties:
        breakfastMenu:
          type: string
        breakfastTime:
          type: string
        lunchMenu:
          type: string
        dinnerMenu:
          type: string
        lunchTime:
          type: string
        dinnerTime:
          type: string
        expectedPeople:
          type: integer
        notes:
          type: string
        status:
          type: string
          enum: [draft, published]
      required:
        - lunchMenu
        - dinnerMenu
        - lunchTime
        - dinnerTime
        - expectedPeople
        - status
    CreateMealBody:
      type: object
      properties:
        menu:
          type: string
        expectedPeople:
          type: integer
      required:
        - menu
        - expectedPeople
    MealPrediction:
      type: object
      properties:
        predictedMeals:
          type: integer
        confidence:
          type: number
        message:
          type: string
      required:
        - predictedMeals
        - confidence
        - message
    LeftoverBody:
      type: object
      properties:
        leftoverMeals:
          type: integer
        actualServed:
          type: integer
      required:
        - leftoverMeals
        - actualServed
    NgoRequest:
      type: object
      properties:
        id:
          type: integer
        pgName:
          type: string
        pgLocation:
          type: string
        availableMeals:
          type: integer
        pickupTime:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected, completed]
        createdAt:
          type: string
        mealMenu:
          type: string
      required:
        - id
        - pgName
        - pgLocation
        - availableMeals
        - pickupTime
        - status
        - createdAt
...