# lib/api-zod/src/generated/types/meal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MealSync API
 * OpenAPI spec version: 0.1.0
 */
import type { MealStatus } from "./mealStatus";

export interface Meal {
  id: number;
  menu?: string;
  breakfastMenu?: string;
  breakfastTime?: string;
  lunchMenu?: string;
  dinnerMenu?: string;
  lunchTime?: string;
  dinnerTime?: string;
  notes?: string;
  expectedPeople: number;
  predictedMeals?: number;
  actualServed?: number;
  leftoverMeals?: number;
  date: string;
  ngoNotified?: boolean;
  status: MealStatus;
}