# lib/api-zod/src/generated/types/upsertMenuBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MealSync API
 * OpenAPI spec version: 0.1.0
 */
import type { UpsertMenuBodyStatus } from "./upsertMenuBodyStatus";

export interface UpsertMenuBody {
  breakfastMenu?: string;
  breakfastTime?: string;
  lunchMenu: string;
  dinnerMenu: string;
  lunchTime: string;
  dinnerTime: string;
  expectedPeople: number;
  notes?: string;
  status: UpsertMenuBodyStatus;
}